We have added a transaction_identifier field to payments
See changelog 2019-11-19: https://developer.moneybird.com/changelog/

We have added a transaction_identifier field to payments. This allows you to store the identifier of a Payment Service Provider transaction. If you use Mollie for payments, the identifier will be used to automatically link the payments to the pay out of Mollie on your bank account.